watch_bspwm: Stop warning upon focused urgent desktops
I think I did this because (1) I thought it wouldn't (and shouldn't)
occur, and (2) I didn't know / hadn't decided what to do if it *did*
happen for some reason.

Turns out this happens often enough that it's annoying.  It also
doesn't really matter since it immediately becomes non-urgent again
because it's focused.